+++ src/subtitles/Rasterizer.cpp
@@ -1736,11 +1736,11 @@ SharedPtrByte Rasterizer::CompositeAlphaMask(const SharedPtrOverlay& overlay, co
         while( last_x<w+xo )
         {   
             byte alpha = sw[0]>>24; 
-            while( sw[3]<w+xo && (sw[2]>>24)==alpha )
+            while( sw[3]<(DWORD)(w+xo) && (sw[2]>>24)==alpha )
             {
                 sw += 2;
             }
-            int new_x = sw[3] < w+xo ? sw[3] : w+xo;
+            int new_x = sw[3] < (DWORD)(w+xo) ? sw[3] : w+xo;
             overlay->FillAlphaMash(s_base, fBody, fBorder, 
                 last_x, yo, new_x-last_x, h, 
                 alpha_mask_data, alpha_mask_pitch,
@@ -1814,7 +1814,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
     // Every remaining line in the bitmap to be rendered...
     switch(draw_method)
     {
-    case   DM::SINGLE_COLOR |   DM::SSE2 | 0*DM::AYUV_PLANAR :
+    case DM::SINGLE_COLOR | DM::SSE2 | 0*DM::AYUV_PLANAR :
     {
         while(h--)
         {
@@ -1828,7 +1828,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
         }
     }
     break;
-    case   DM::SINGLE_COLOR | 0*DM::SSE2 | 0*DM::AYUV_PLANAR :
+    case DM::SINGLE_COLOR | 0*DM::SSE2 | 0*DM::AYUV_PLANAR :
     {
         while(h--)
         {
@@ -1839,7 +1839,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
         }
     }
     break;
-    case 0*DM::SINGLE_COLOR |   DM::SSE2 | 0*DM::AYUV_PLANAR :
+    case 0*DM::SINGLE_COLOR | DM::SSE2 | 0*DM::AYUV_PLANAR :
     {
         while(h--)
         {
@@ -1849,7 +1849,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
                 // xo is the offset (usually negative) we have moved into the image
                 // So if we have passed the switchpoint (?) switch to another colour
                 // (So switchpts stores both colours *and* coordinates?)
-                if(wt+xo >= sw[1]) {while(wt+xo >= sw[1]) sw += 2; color = sw[-2];}
+                if((DWORD)(wt+xo) >= sw[1]) {while((DWORD)(wt+xo) >= sw[1]) sw += 2; color = sw[-2];}
                 pixmix_sse2(&dst[wt], color, s[wt]);
             }
             s += overlayPitch;
@@ -1864,7 +1864,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
             const DWORD *sw = switchpts;
             for(int wt=0; wt<w; ++wt)
             {
-                if(wt+xo >= sw[1]) {while(wt+xo >= sw[1]) sw += 2; color = sw[-2];}
+                if((DWORD)(wt+xo) >= sw[1]) {while((DWORD)(wt+xo) >= sw[1]) sw += 2; color = sw[-2];}
                 pixmix(&dst[wt], color, s[wt]);
             }
             s += overlayPitch;
@@ -1872,7 +1872,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
         }
     }
     break;
-    case   DM::SINGLE_COLOR |   DM::SSE2 |   DM::AYUV_PLANAR :
+    case DM::SINGLE_COLOR | DM::SSE2 | DM::AYUV_PLANAR :
     {
         unsigned char* dst_A = bitmap->plans[0] + dst_offset;
         unsigned char* dst_Y = bitmap->plans[1] + dst_offset;
@@ -1885,7 +1885,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
         AlphaBlt(dst_A, s, 0, h, w, overlayPitch, bitmap->pitch);
     }
     break;
-    case 0*DM::SINGLE_COLOR |   DM::SSE2 |   DM::AYUV_PLANAR :
+    case 0*DM::SINGLE_COLOR | DM::SSE2 | DM::AYUV_PLANAR :
     {
         unsigned char* dst_A = bitmap->plans[0] + dst_offset;
         unsigned char* dst_Y = bitmap->plans[1] + dst_offset;
@@ -1897,7 +1897,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
         color = sw[0];
         while(last_x<w+xo)
         {
-            int new_x = sw[3] < w+xo ? sw[3] : w+xo;
+            int new_x = sw[3] < (DWORD)(w+xo) ? sw[3] : w+xo;
             color = sw[0];
             sw += 2;
             if( new_x < last_x )
@@ -1915,7 +1915,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
         }
     }
     break;
-    case   DM::SINGLE_COLOR | 0*DM::SSE2 |   DM::AYUV_PLANAR :
+    case DM::SINGLE_COLOR | 0*DM::SSE2 | DM::AYUV_PLANAR :
     {
 //        char * debug_dst=(char*)dst;int h2 = h;
 //        XY_DO_ONCE( xy_logger::write_file("G:\\b2_rt", (char*)&color, sizeof(color)) );
@@ -1955,7 +1955,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
 //        XY_DO_ONCE( xy_logger::write_file("G:\\a2_rt", debug_dst, (h2-1)*spd.pitch) );
     }
     break;
-    case 0*DM::SINGLE_COLOR | 0*DM::SSE2 |   DM::AYUV_PLANAR :
+    case 0*DM::SINGLE_COLOR | 0*DM::SSE2 | DM::AYUV_PLANAR :
     {
         unsigned char* dst_A = bitmap->plans[0] + dst_offset;
         unsigned char* dst_Y = bitmap->plans[1] + dst_offset;
@@ -1966,7 +1966,7 @@ void Rasterizer::Draw(XyBitmap* bitmap, SharedPtrOverlay overlay, const CRect& c
             const DWORD *sw = switchpts;
             for(int wt=0; wt<w; ++wt)
             {
-                if(wt+xo >= sw[1]) {while(wt+xo >= sw[1]) sw += 2; color = sw[-2];}
+                if((DWORD)(wt+xo) >= sw[1]) {while((DWORD)(wt+xo) >= sw[1]) sw += 2; color = sw[-2];}
                 DWORD temp = COMBINE_AYUV(dst_A[wt], dst_Y[wt], dst_U[wt], dst_V[wt]);
                 pixmix(&temp, color, (s[wt]*(color>>24))>>8);
                 SPLIT_AYUV(temp, dst_A+wt, dst_Y+wt, dst_U+wt, dst_V+wt);
@@ -1995,7 +1995,7 @@ void Rasterizer::FillSolidRect(SubPicDesc& spd, int x, int y, int nWidth, int nH
 
     switch (draw_method)
     {
-    case   DM::SSE2 | 0*DM::AYUV_PLANAR :
+    case DM::SSE2 | 0*DM::AYUV_PLANAR :
     {
         for (int wy=y; wy<y+nHeight; wy++) {
             DWORD* dst = (DWORD*)((BYTE*)spd.bits + spd.pitch * wy) + x;
@@ -2015,7 +2015,7 @@ void Rasterizer::FillSolidRect(SubPicDesc& spd, int x, int y, int nWidth, int nH
         }
     }
     break;
-    case   DM::SSE2 |   DM::AYUV_PLANAR :
+    case DM::SSE2 | DM::AYUV_PLANAR :
     {
         BYTE* dst = reinterpret_cast<BYTE*>(spd.bits) + spd.pitch * y + x;
         BYTE* dst_A = dst;
@@ -2028,7 +2028,7 @@ void Rasterizer::FillSolidRect(SubPicDesc& spd, int x, int y, int nWidth, int nH
         AlphaBlt(dst_A, argb>>24, 0, nHeight, nWidth, spd.pitch);
     }
     break;
-    case 0*DM::SSE2 |   DM::AYUV_PLANAR :
+    case 0*DM::SSE2 | DM::AYUV_PLANAR :
     {
         BYTE* dst = reinterpret_cast<BYTE*>(spd.bits) + spd.pitch * y + x;
         BYTE* dst_A = dst;
